conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.6.5=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=20.1.0=py38h27cfd23_1
async-timeout=4.0.2=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
audioread=2.1.9=py38h578d9bd_2
autopage=0.5.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pypi_0
cliff=3.10.0=pypi_0
cloudpickle=2.0.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.3.3=pypi_0
colorlog=6.6.0=pypi_0
cryptography=36.0.0=py38h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
databricks-cli=0.16.4=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker=5.0.3=pypi_0
entrypoints=0.3=py38_0
expat=2.4.4=h295c915_2
ffmpeg=4.2.2=h20bf706_0
flask=2.0.2=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=pypi_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.26=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.43.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gunicorn=20.1.0=pypi_0
hydra-core=1.1.1=pypi_0
hydra-optuna-sweeper=1.1.2=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py38h06a4308_1
ipython=7.31.1=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.0.1=pypi_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libllvm11=11.1.0=h3826bc1_0
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
librosa=0.9.0=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.37.0=py38h295c915_1
lz4-c=1.9.3=h295c915_1
mako=1.1.6=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mlflow=1.23.1=pypi_0
multidict=6.0.2=pypi_0
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.3.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
nnaudio=0.3.1=pypi_0
notebook=6.4.6=py38h06a4308_0
numba=0.54.1=py38h51133e4_0
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.1.1=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
optuna=2.10.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.0=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pbr=5.8.1=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prettytable=3.0.0=pypi_0
prometheus-flask-exporter=0.18.7=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyloudnorm=0.1.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
python_abi=3.8=2_cp38
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.2.2=pyhd3eb1b0_0
qtpy=1.11.2=pyhd3eb1b0_0
querystring-parser=1.2.4=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.2.2=py_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=py38h51133e4_1
scipy=1.7.3=py38hc147768_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=59.5.0=pypi_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sqlalchemy=1.4.31=pypi_0
sqlite=3.37.2=hc218d9a_0
sqlparse=0.4.2=pypi_0
stevedore=3.5.0=pypi_0
tabulate=0.8.9=pypi_0
tbb=2021.5.0=hd09550d_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.1=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.2=py38_cu113
torchcrepe=0.0.15=pypi_0
torchmetrics=0.7.2=pypi_0
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.2.3=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
